+++ src/Repositories/HostnameRepository.php
@@ -22,6 +22,7 @@ use Hyn\Tenancy\Traits\DispatchesEvents;
 use Hyn\Tenancy\Validators\HostnameValidator;
 use Illuminate\Contracts\Cache\Factory;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Arr;
 
 class HostnameRepository implements Contract
 {
@@ -56,7 +57,7 @@ class HostnameRepository implements Contract
      * @param string $hostname
      * @return Hostname|null
      */
-    public function findByHostname(string $hostname): ?Hostname
+    public function findByHostname(string $hostname)
     {
         return $this->cache->remember("tenancy.hostname.$hostname", config('tenancy.hostname.cache'), function () use ($hostname) {
             return $this->hostname->newQuery()->where('fqdn', $hostname)->first();
@@ -66,7 +67,7 @@ class HostnameRepository implements Contract
     /**
      * @return Hostname|null
      */
-    public function getDefault() : ?Hostname
+    public function getDefault()
     {
         if (config('tenancy.hostname.default')) {
             return $this->hostname->newQuery()->where('fqdn', config('tenancy.hostname.default'))->first();
@@ -116,14 +117,20 @@ class HostnameRepository implements Contract
 
         $this->validator->save($hostname);
 
-        $dirty = $hostname->getDirty();
+        $dirty = collect(array_keys($hostname->getDirty()))->mapWithKeys(function ($value, $key) use ($hostname) {
+            return [ $value => $hostname->getOriginal($value) ];
+        });
 
         $hostname->save();
 
         $this->cache->forget("tenancy.hostname.{$hostname->fqdn}");
 
+        if ($dirty->has('fqdn')) {
+            $this->cache->forget("tenancy.hostname.{$dirty->get('fqdn')}");
+        }
+
         $this->emitEvent(
-            new Events\Updated($hostname, $dirty)
+            new Events\Updated($hostname, $dirty->toArray())
         );
 
         return $hostname;