+++ changelog.md
@@ -1,3 +1,21 @@
-- 3.0.0
-	- Created tenancy:migrate command.
-	- Complete rewrite from scratch.
+- 5.0.5
+    - Moved to CircleCI for CI, now fully separated from GitLab.
+    - Improved validation of hostnames by allowing nullables.
+    - #284, fixed validation issue when updating hostname or website.
+    - Fixed maintenance mode in HostnameActions middleware.
+    
+- 5.0.4
+    - Flushing cache whenever hostname fqdn or website uuid changes.
+- 5.0.3
+    - Added seed command and auto seeding of newly auto-migrated tenants.
+    - Added more strict validation.
+    - Added script to clean tenant and tenancy databases while running local tests.
+    - Fixed an issue with nginx where the media folder was incorrectly mapped.
+    - Fixed issue where tenant hostname was identified using the wrong header.
+- 5.0.2
+    - #262, allows using the tenant disk.
+    - #260, triggering events in case tenant directories are mutated.
+- 5.0.1
+    - Added initial Nginx support.
+    - Added MySQL database support, set tenancy > website > uuid-limit-length-to-32.
+- 5.0.0